data access layer + settings
