+++ .idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="..\:/Users/Epikg/Documents/MovieDisplayApplication/app/src/main/res/layout/bookmarked_repos.xml" value="0.14722222222222223" />
         <entry key="..\:/Users/Epikg/Documents/MovieDisplayApplication/app/src/main/res/layout/movie_detail.xml" value="0.14722222222222223" />
+        <entry key="..\:/Users/bujbu/Google Drive/repos/MovieDisplayApplication/app/src/main/res/xml/settings.xml" value="0.18206521739130435" />
         <entry key="app/src/main/res/layout/activity_bookmarked_repos.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.1503623188405797" />
         <entry key="app/src/main/res/layout/activity_repo_detail.xml" value="0.278125" />
